api/requirements.txt
fastapi>=0.109.2
pydantic>=2.6.1
python-multipart>=0.0.9
google-generativeai>=0.8.0
